Build 2197 Release Date: 11 May 2012

+++ Default/side_bar.py
@@ -33,11 +33,13 @@ class NewFolderCommand(sublime_plugin.WindowCommand):
 
 class DeleteFolderCommand(sublime_plugin.WindowCommand):
     def run(self, dirs):
-        try:
-            for d in dirs:
-                send2trash.send2trash(d)
-        except:
-            sublime.status_message("Unable to delete folder")
+
+        if sublime.ok_cancel_dialog("Delete Folder?", "Delete"):
+            try:
+                for d in dirs:
+                    send2trash.send2trash(d)
+            except:
+                sublime.status_message("Unable to delete folder")
 
     def is_visible(self, dirs):
         return len(dirs) > 0
@@ -45,13 +47,21 @@ class DeleteFolderCommand(sublime_plugin.WindowCommand):
 class RenamePathCommand(sublime_plugin.WindowCommand):
     def run(self, paths):
         branch, leaf = os.path.split(paths[0])
-        self.window.show_input_panel("New Name:", leaf, functools.partial(self.on_done, paths[0], branch), None, None)
+        v = self.window.show_input_panel("New Name:", leaf, functools.partial(self.on_done, paths[0], branch), None, None)
+        name, ext = os.path.splitext(leaf)
+
+        v.sel().clear()
+        v.sel().add(sublime.Region(0, len(name)))
 
     def on_done(self, old, branch, leaf):
         new = os.path.join(branch, leaf)
 
         try:
             os.rename(old, new)
+
+            v = self.window.find_open_file(old)
+            if v:
+                v.retarget(new)
         except:
             sublime.status_message("Unable to rename")
